xkcp_sys

Function Keccak_HashInitialize

pub unsafe extern "C" fn Keccak_HashInitialize(
    hashInstance: *mut Keccak_HashInstance,
    rate: c_uint,
    capacity: c_uint,
    hashbitlen: c_uint,
    delimitedSuffix: c_uchar,
) -> HashReturn

Function to initialize the Keccak[r, c] sponge function instance used in sequential hashing mode. @param hashInstance Pointer to the hash instance to be initialized. @param rate The value of the rate r. @param capacity The value of the capacity c. @param hashbitlen The desired number of output bits, or 0 for an arbitrarily-long output. @param delimitedSuffix Bits that will be automatically appended to the end of the input message, as in domain separation. This is a byte containing from 0 to 7 bits formatted like the @a delimitedData parameter of the Keccak_SpongeAbsorbLastFewBits() function. @pre One must have r+c=1600 and the rate a multiple of 8 bits in this implementation. @return KECCAK_SUCCESS if successful, KECCAK_FAIL otherwise.