xkcp_sys

Function Keccak_HashUpdate

pub unsafe extern "C" fn Keccak_HashUpdate(
    hashInstance: *mut Keccak_HashInstance,
    data: *const BitSequence,
    databitlen: BitLength,
) -> HashReturn

Function to give input data to be absorbed. @param hashInstance Pointer to the hash instance initialized by Keccak_HashInitialize(). @param data Pointer to the input data. When @a databitLen is not a multiple of 8, the last bits of data must be in the least significant bits of the last byte (little-endian convention). In this case, the (8 - @a databitLen mod 8) most significant bits of the last byte are ignored. @param databitLen The number of input bits provided in the input data. @pre In the previous call to Keccak_HashUpdate(), databitlen was a multiple of 8. @return KECCAK_SUCCESS if successful, KECCAK_FAIL otherwise.