xkcp_sys

Function Kravatte_SANSE_Unwrap

pub unsafe extern "C" fn Kravatte_SANSE_Unwrap(
    kvInstance: *mut Kravatte_SANSE_Instance,
    ciphertext: *const BitSequence,
    plaintext: *mut BitSequence,
    dataBitLen: BitLength,
    AD: *const BitSequence,
    ADBitLen: BitLength,
    tag: *const c_uchar,
) -> c_int

Function to unwrap ciphertext into plaintext. @param kvInstance Pointer to the instance initialized by Kravatte_SANSE_Initialize(). @param ciphertext Pointer to ciphertext data to unwrap. @param plaintext Pointer to buffer where the full unwrapped data will be stored. The plaintext buffer must not overlap ciphertext. @param dataBitLen The size of the ciphertext/plaintext data. @param AD Pointer to the Associated Data. @param ADBitLen The number of bytes provided in the Associated Data. @param tag The buffer where to read the tag to check (when lastFlag is set). This buffer must be minimum Kravatte_SANSE_TagLength bytes long. @return 0 if successful, 1 otherwise.