Function xkcp_sys::Vatte

pub unsafe extern "C" fn Vatte(
    kvInstance: *mut Kravatte_Instance,
    output: *mut BitSequence,
    outputBitLen: BitLength,
    flags: c_int
) -> c_int

Function to expand output data. @param kvInstance Pointer to the hash instance initialized by Kravatte_MaskDerivation(). @param output Pointer to the buffer where to store the output data. @param outputBitLen The number of output bits desired. This must be a multiple of 8 if KRAVATTE_FLAG_LAST_PART flag not set. @param flags Bitwise or combination of KRAVATTE_FLAG_NONE, KRAVATTE_FLAG_SHORT, KRAVATTE_FLAG_LAST_PART. @return 0 if successful, 1 otherwise.